Renamed package to eventhub
