+++ azure/eventhub/tests/utils.py
@@ -8,7 +8,7 @@ import logging
 from logging.handlers import RotatingFileHandler
 
 def get_logger(filename, level=logging.INFO):
-    logger = logging.getLogger("eventhubs")
+    logger = logging.getLogger("eventhub")
     logger.setLevel(level)
     formatter = logging.Formatter('%(asctime)s %(name)-12s %(levelname)-8s %(message)s')
     file_handler = RotatingFileHandler(filename, maxBytes=5*1024*1024, backupCount=2)
