+++ examples/recv_batch.py
@@ -13,7 +13,7 @@ the event in on_event_data callback.
 import os
 import sys
 import logging
-from eventhubs import EventHubClient, Receiver, Offset
+from azure.eventhub import EventHubClient, Receiver, Offset
 
 import examples
 logger = examples.get_logger(logging.INFO)
